Updated to new testing framework.
