+++ .github/workflows/redcap-tests.yml
@@ -1,6 +1,6 @@
-# Run selenium tests on a REDCap External Module
+# Run selenium tests on REDCap
 
-name: REDCap External Module Tests
+name: REDCap Tests
 
 on:
   pull_request:
@@ -11,11 +11,11 @@ on:
 jobs:
   jobs:
     name: Jobs
-    uses: Nottingham-CTU/Workflows/.github/workflows/redcap-module-tests.yml@main
+    uses: Nottingham-CTU/Workflows/.github/workflows/redcap-tests.yml@main
     with:
       RC_COMMUNITY_USERNAME: ${{ vars.RC_COMMUNITY_USERNAME }}
       RC_INSTALLED_VERSION: ${{ vars.RC_INSTALLED_VERSION }}
       USERS_NAMES: ${{ vars.USERS_NAMES }}
     secrets:
       RC_COMMUNITY_PASSWORD: ${{ secrets.RC_COMMUNITY_PASSWORD }}
-
+      SECRET1: ${{ secrets.SYSDOCS_FILE_UPDATE_URL }}
