+++ tests/details.json
@@ -1,92 +1,77 @@
 [
 	{
-		"file" : "test_t01Createcategoryviewpacks.py",
+		"file" : "test_01_Create_category_view_packs.py",
 		"name" : "Create category / view packs",
-		"expected" : "Pack category can be created and packs can be viewed only by admins and users with viewing roles",
-		"scenario" : "Create pack category, define roles for management features, confirm only admins and users with viewing roles can view the pack list"
+		"expected" : "Pack category can be created and packs can be viewed only by admins and users with viewing roles"
 	},
 	{
-		"file" : "test_t02Addpacks.py",
+		"file" : "test_02_Add_packs.py",
 		"name" : "Add packs",
-		"expected" : "Packs can be added to a pack category only by admins and users with adding roles",
-		"scenario" : "Create pack category, define roles and confirm only admins and users with adding roles can add new packs"
+		"expected" : "Packs can be added to a pack category only by admins and users with adding roles"
 	},
 	{
-		"file" : "test_t03AssignpackformsubmitDAG.py",
+		"file" : "test_03_Assign_pack_form_submit_DAG.py",
 		"name" : "Assign pack on form submit (w/DAG)",
-		"expected" : "Packs are assigned to record only when issued to the respective DAG",
-		"scenario" : "Create pack category and confirm packs assigned only when applicable"
+		"expected" : "Packs are assigned to record only when issued to the respective DAG"
 	},
 	{
-		"file" : "test_t04Assignpackformsubmit.py",
+		"file" : "test_04_Assign_pack_form_submit.py",
 		"name" : "Assign pack on form submit",
-		"expected" : "Packs are assigned once added",
-		"scenario" : "Create pack category and confirm packs assigned only when applicable"
+		"expected" : "Packs are assigned once added"
 	},
 	{
-		"file" : "test_t05Assignpackformsubmitexpiry.py",
+		"file" : "test_05_Assign_pack_form_submit_expiry.py",
 		"name" : "Assign pack on form submit (w/expiry)",
-		"expected" : "Packs are assigned only if they are not expired",
-		"scenario" : "Create pack category and confirm packs assigned only when applicable"
+		"expected" : "Packs are assigned only if they are not expired"
 	},
 	{
-		"file" : "test_t06Assignpackformsubmitinvalid.py",
+		"file" : "test_06_Assign_pack_form_submit_invalid.py",
 		"name" : "Assign pack on form submit (w/invalid)",
-		"expected" : "Packs are assigned only if they are not invalid",
-		"scenario" : "Create pack category and confirm packs assigned only when applicable"
+		"expected" : "Packs are assigned only if they are not invalid"
 	},
 	{
-		"file" : "test_t07IssuepackstoDAG.py",
+		"file" : "test_07_Issue_packs_to_DAG.py",
 		"name" : "Issue packs to DAG",
-		"expected" : "Packs can be issued to a DAG only by admins and users with issuing roles",
-		"scenario" : "Create pack category, define roles and confirm only admins and users with issuing roles can issue packs to DAGs"
+		"expected" : "Packs can be issued to a DAG only by admins and users with issuing roles"
 	},
 	{
-		"file" : "test_t08Markpacksinvalid.py",
+		"file" : "test_08_Mark_packs_invalid.py",
 		"name" : "Mark packs as invalid",
-		"expected" : "Packs can be marked invalid only by admins and users with roles for marking invalid",
-		"scenario" : "Create pack category, define roles and confirm only admins and users with roles for marking invalid can mark packs as invalid"
+		"expected" : "Packs can be marked invalid only by admins and users with roles for marking invalid"
 	},
 	{
-		"file" : "test_t09Assignpackminimization.py",
+		"file" : "test_09_Assign_pack_minimization.py",
 		"name" : "Assign pack on minimization",
-		"expected" : "Packs are assigned when requested by the Minimization module, provided the requirements are met",
-		"scenario" : "Create pack category and check the assignment takes place, or not, depending on minimized allocation and settings for handling lack of minimized pack"
+		"expected" : "Packs are assigned when requested by the Minimization module, provided the requirements are met"
 	},
 	{
-		"file" : "test_t10Assignpackautomatic.py",
+		"file" : "test_10_Assign_pack_automatic.py",
 		"name" : "Assign pack automatically",
-		"expected" : "Packs are assigned by execution of the cron function",
-		"scenario" : "Create pack category and check the assignment takes place as expected"
+		"expected" : "Packs are assigned by execution of the cron function"
 	},
 	{
-		"file" : "test_t11Packexpirybuffer.py",
+		"file" : "test_11_Pack_expiry_buffer.py",
 		"name" : "Pack expiry buffer",
-		"expected" : "Packs which have not expired but are closer to expiry than the buffer are treated as expired and are not assigned",
-		"scenario" : "Create pack category with an expiry buffer, add a pack which expires within the buffer and ensure it is not assigned. Then reduce the expiry buffer so the pack does not expire within the buffer and ensure it can now be assigned."
+		"expected" : "Packs which have not expired but are closer to expiry than the buffer are treated as expired and are not assigned"
 	},
 	{
-		"file" : "test_t12Manualreassignment.py",
+		"file" : "test_12_Manual_reassignment.py",
 		"name" : "Manual reassignment",
-		"expected" : "Packs can be manually assigned, unassigned, reassigned and exchanged by users with the manual reassignment roles",
-		"scenario" : "Create pack category, define roles and confirm packs can be manually assigned, unassigned, reassigned and exchanged (and that pack fields are swapped out in the record data as expected) only by admins and users with roles for manual reassignemnt."
+		"expected" : "Packs can be manually assigned, unassigned, reassigned and exchanged by users with the manual reassignment roles"
 	},
 	{
-		"file" : "test_t13Editanddeletepacks.py",
+		"file" : "test_13_Edit_and_delete_packs.py",
 		"name" : "Edit and delete packs",
-		"expected" : "Packs can be edited and deleted by users with the edit and delete roles",
-		"scenario" : "Create pack category, define roles and confirm packs can be edited and deleted only by admins and users with roles for editing and deleting."
+		"expected" : "Packs can be edited and deleted by users with the edit and delete roles"
 	},
 	{
-		"file" : "test_t14Assignpackselection.py",
+		"file" : "test_14_Assign_pack_selection.py",
 		"name" : "Assign pack using selection trigger",
-		"expected" : "Packs are visible in a drop down once added and can be selected to trigger assignment",
-		"scenario" : "Create pack category and confirm packs assigned only when applicable"
+		"expected" : "Packs are visible in a drop down once added and can be selected to trigger assignment"
 	},
 	{
-		"file" : "test_t15AssignpackselectionDAG.py",
+		"file" : "test_15_Assign_pack_selection_DAG.py",
 		"name" : "Assign pack using selection trigger (w/DAG)",
-		"expected" : "Packs are visible in a drop down only when issued to the respective DAG and can be selected to trigger assignment",
-		"scenario" : "Create pack category and confirm packs assigned only when applicable"
+		"expected" : "Packs are visible in a drop down only when issued to the respective DAG and can be selected to trigger assignment"
 	}
 ]
