+++ tests/init.py
@@ -1,32 +1,29 @@
-# Generated by Selenium IDE
+# Generated from Selenium IDE
+# Test name: init
 import pytest
 import time
 import json
 from selenium import webdriver
-from selenium.webdriver.common.by import By
 from selenium.webdriver.common.action_chains import ActionChains
+from selenium.webdriver.common.by import By
+from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 
-class TestInit():
+class Test_init:
   def setup_method(self, method):
-    self.driver = webdriver.Firefox()
+    self.driver = self.selectedBrowser
     self.vars = {}
-  
   def teardown_method(self, method):
     self.driver.quit()
-  
+
   def test_init(self):
     self.driver.get("http://127.0.0.1/")
     self.driver.find_element(By.LINK_TEXT, "My Projects").click()
-    elements = self.driver.find_elements(By.XPATH, "//*[@id=\"table-proj_table\"][contains(.,\'Pack Management Test\')]")
-    assert len(elements) == 0
+    assert len(self.driver.find_elements(By.XPATH, "//*[@id=\"table-proj_table\"][contains(.,'Pack Management Test')]")) == 0
     self.driver.find_element(By.LINK_TEXT, "New Project").click()
     self.driver.find_element(By.ID, "app_title").send_keys("Pack Management Test")
-    dropdown = self.driver.find_element(By.ID, "purpose")
-    dropdown.find_element(By.XPATH, "//option[. = 'Practice / Just for fun']").click()
+    self.driver.find_element(By.ID, "purpose").find_element(By.XPATH, "(descendant::option)[. = 'Practice / Just for fun']").click()
     self.driver.find_element(By.ID, "project_template_radio1").click()
     self.driver.find_element(By.CSS_SELECTOR, "input[name=\"copyof\"][value=\"2\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, ".btn-primaryrc").click()
@@ -35,77 +32,70 @@ class TestInit():
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "tr[data-module=\"pack_management\"] .enable-button")))
     self.driver.find_element(By.CSS_SELECTOR, "tr[data-module=\"pack_management\"] .enable-button").click()
     time.sleep(5)
-    elements = self.driver.find_elements(By.CSS_SELECTOR, "#external-modules-enabled tr[data-module=\"pack_management\"]")
-    assert len(elements) > 0
+    assert len(self.driver.find_elements(By.CSS_SELECTOR, "#external-modules-enabled tr[data-module=\"pack_management\"]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"Design/online_designer.php\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=visit_lab_data\"]").click()
     self.driver.find_element(By.ID, "btn-last").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "field_type")))
-    dropdown = self.driver.find_element(By.ID, "field_type")
-    dropdown.find_element(By.CSS_SELECTOR, "*[value='text']").click()
+    self.driver.find_element(By.ID, "field_type").find_element(By.CSS_SELECTOR, "*[value='text']").click()
     self.driver.find_element(By.ID, "field_label").send_keys("Pack")
     self.driver.find_element(By.ID, "field_name").send_keys("pack_id")
-    self.driver.find_element(By.XPATH, "//button[contains(.,\'Save\')]").click()
+    self.driver.find_element(By.XPATH, "//button[contains(.,'Save')]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "design-pack_id")))
     self.driver.find_element(By.ID, "btn-last").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "field_type")))
-    dropdown = self.driver.find_element(By.ID, "field_type")
-    dropdown.find_element(By.CSS_SELECTOR, "*[value='text']").click()
+    self.driver.find_element(By.ID, "field_type").find_element(By.CSS_SELECTOR, "*[value='text']").click()
     self.driver.find_element(By.ID, "field_label").send_keys("Pack count")
     self.driver.find_element(By.ID, "field_name").send_keys("pack_count")
-    self.driver.find_element(By.XPATH, "//button[contains(.,\'Save\')]").click()
+    self.driver.find_element(By.XPATH, "//button[contains(.,'Save')]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "design-pack_count")))
     self.driver.find_element(By.ID, "btn-last").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "field_type")))
-    dropdown = self.driver.find_element(By.ID, "field_type")
-    dropdown.find_element(By.CSS_SELECTOR, "*[value='text']").click()
+    self.driver.find_element(By.ID, "field_type").find_element(By.CSS_SELECTOR, "*[value='text']").click()
     self.driver.find_element(By.ID, "field_label").send_keys("Pack assign date")
     self.driver.find_element(By.ID, "field_name").send_keys("pack_date")
-    dropdown = self.driver.find_element(By.ID, "val_type")
-    dropdown.find_element(By.CSS_SELECTOR, "*[value='datetime_dmy']").click()
-    self.driver.find_element(By.XPATH, "//button[contains(.,\'Save\')]").click()
+    self.driver.find_element(By.ID, "val_type").find_element(By.CSS_SELECTOR, "*[value='datetime_dmy']").click()
+    self.driver.find_element(By.XPATH, "//button[contains(.,'Save')]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "design-pack_date")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"route=DataAccessGroupsController:index\"]").click()
     self.driver.find_element(By.ID, "new_group").send_keys("DAG1")
     self.driver.find_element(By.ID, "new_group_button").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//td[contains(.,\'DAG1\')]")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//td[contains(.,'DAG1')]")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"DataEntry/record_status_dashboard.php\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "button[onclick*=\"DataEntry/record_home.php\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"DataEntry/index.php\"][href*=\"page=demographics\"]").click()
-    dropdown = self.driver.find_element(By.NAME, "__GROUPID__")
-    dropdown.find_element(By.XPATH, "//option[. = 'DAG1']").click()
-    self.driver.execute_script("autoFill();$(\'#south\').remove()")
-    self.driver.execute_script("dataEntrySubmit(\'submit-btn-saverecord\')")
+    self.driver.find_element(By.NAME, "__GROUPID__").find_element(By.XPATH, "(descendant::option)[. = 'DAG1']").click()
+    self.driver.execute_script("autoFill();$('#south').remove()")
+    self.driver.execute_script("dataEntrySubmit('submit-btn-saverecord')")
     WebDriverWait(self.driver, 60).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"UserRights/index.php\"]").click()
     self.driver.find_element(By.ID, "new_rolename").send_keys("PackView")
     self.driver.find_element(By.ID, "createRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]")))
     self.driver.find_element(By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id=\'user_rights_roles_table\']//td[contains(.,\'PackView\')]")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td[contains(.,'PackView')]")))
     self.driver.find_element(By.ID, "new_rolename").send_keys("PackIssue")
     self.driver.find_element(By.ID, "createRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]")))
     self.driver.find_element(By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id=\'user_rights_roles_table\']//td[contains(.,\'PackIssue\')]")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td[contains(.,'PackIssue')]")))
     self.driver.find_element(By.ID, "new_rolename").send_keys("PackInvalid")
     self.driver.find_element(By.ID, "createRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]")))
     self.driver.find_element(By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id=\'user_rights_roles_table\']//td[contains(.,\'PackInvalid\')]")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td[contains(.,'PackInvalid')]")))
     self.driver.find_element(By.ID, "new_rolename").send_keys("PackAssign")
     self.driver.find_element(By.ID, "createRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]")))
     self.driver.find_element(By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id=\'user_rights_roles_table\']//td[contains(.,\'PackAssign\')]")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td[contains(.,'PackAssign')]")))
     self.driver.find_element(By.ID, "new_rolename").send_keys("PackAdd")
     self.driver.find_element(By.ID, "createRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]")))
     self.driver.find_element(By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id=\'user_rights_roles_table\']//td[contains(.,\'PackAdd\')]")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td[contains(.,'PackAdd')]")))
     self.driver.find_element(By.ID, "new_rolename").send_keys("PackEdit")
     self.driver.find_element(By.ID, "createRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]")))
     self.driver.find_element(By.CSS_SELECTOR, ".ui-dialog button[style*=\"bold\"]").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id=\'user_rights_roles_table\']//td[contains(.,\'PackEdit\')]")))
-  
+    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td[contains(.,'PackEdit')]")))
