+++ tests/output_end.html
@@ -1,0 +1,6 @@
+     </td>
+    </tr>
+   </tbody>
+  </table>
+ </body>
+</html>
