+++ tests/test_02_Add_packs.py
@@ -1,6 +1,5 @@
 # Generated from Selenium IDE
 # Test name: t02 Add packs
-# Comment: Check that packs can be added to a new pack category - admin and users with appropriate role should be able to do this
 import pytest
 import time
 import json
@@ -10,6 +9,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
+from fn_switchuser import Test_fn_switchuser as Sub1
 
 class Test_02_Add_packs:
   def setup_method(self, method):
@@ -20,7 +20,6 @@ class Test_02_Add_packs:
 
   def test_02_Add_packs(self):
     self.driver.get("http://127.0.0.1/")
-    self.driver.execute_script("sessionStorage.setItem('github_js_actions','1')")
     self.driver.find_element(By.LINK_TEXT, "My Projects").click()
     assert len(self.driver.find_elements(By.XPATH, "//*[@id=\"table-proj_table\"][contains(.,'Pack Management Test')]")) > 0
     self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
@@ -30,11 +29,13 @@ class Test_02_Add_packs:
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     time.sleep(2)
     self.driver.execute_script("$('#south').remove()")
-    self.driver.execute_script("$('input[name=\"cat_id\"]').val('packs2')")
+    self.driver.execute_script("$('[name=\"cat_id\"]').css('max-width','200px')")
+    self.driver.find_element(By.NAME, "cat_id").send_keys("packs2")
     self.driver.find_element(By.CSS_SELECTOR, "input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     time.sleep(2)
     self.driver.execute_script("$('#south').remove()")
+    self.driver.execute_script("$('[name=\"logic\"],[name=\"packfield\"],[name=\"datefield\"],[name=\"countfield\"],[name=\"valuefield\"],[name=\"roles_view\"],[name=\"roles_dags\"],[name=\"roles_invalid\"],[name=\"roles_assign\"],[name=\"roles_add\"],[name=\"roles_edit\"]').css('max-width','250px')")
     WebDriverWait(self.driver, 60).until(expected_conditions.presence_of_element_located((By.NAME, "enabled")))
     self.driver.find_element(By.NAME, "enabled").find_element(By.CSS_SELECTOR, "*[value='0']").click()
     self.driver.find_element(By.NAME, "trigger").find_element(By.CSS_SELECTOR, "*[value='F']").click()
@@ -53,29 +54,39 @@ class Test_02_Add_packs:
     self.driver.find_element(By.CSS_SELECTOR, "#catform button.btn-primaryrc").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=configure_edit\"][href*=\"cat_id=packs2\"]")) > 0
+    self.driver.execute_script("//SETDESC:Assert category is saved")
+    self.driver.find_element(By.XPATH, "//td//span[contains(.,'packs2')]").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs2\"]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs2\"]").click()
+    self.driver.execute_script("$('[name=\"id\"],[name=\"value\"]').css('max-width','250px')")
     self.driver.find_element(By.NAME, "id").send_keys("1")
     self.driver.execute_script("$('#south').remove()")
     self.driver.find_element(By.CSS_SELECTOR, "#singlepack input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
+    self.driver.execute_script("$('[name=\"id\"],[name=\"value\"]').css('max-width','250px')")
     assert len(self.driver.find_elements(By.CSS_SELECTOR, ".mod-packmgmt-okmsg")) > 0
+    self.driver.execute_script("//SETDESC:Assert packs added successfully")
+    self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-okmsg").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href=\"#multiplepacks\"]").click()
     self.driver.execute_script("fd=new FormData($('form[enctype=\"multipart/form-data\"]')[0]);fd.set('packs_upload',new Blob(['id\\n2\\n3']));fetch( window.location.href, {body:fd, method:'post'})")
+    self.driver.execute_script("//SAVEDESC:Perform upload of CSV data")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]:not([href*=\"logout=1\"])").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs2\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"pack_id\"][value=\"1\"]")) > 0
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"pack_id\"][value=\"2\"]")) > 0
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"pack_id\"][value=\"3\"]")) > 0
+    self.driver.execute_script("//SETDESC:Assert pack IDs 1, 2 and 3 present")
+    self.driver.find_element(By.CSS_SELECTOR, "input[name=\"pack_id\"][value=\"2\"]").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"UserRights/index.php\"]").click()
-    self.driver.find_element(By.ID, "new_username_assign").send_keys("user")
+    self.driver.find_element(By.ID, "new_username_assign").send_keys("user1")
     self.driver.find_element(By.ID, "assignUserBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "notify_email_role")))
     self.driver.execute_script("$('#notify_email_role').prop('checked',false)")
     self.driver.find_element(By.ID, "user_role").find_element(By.XPATH, "(descendant::option)[. = 'PackAdd']").click()
     self.driver.find_element(By.ID, "assignDagRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td//a[contains(.,'user')]")))
+    time.sleep(3)
     self.driver.find_element(By.ID, "new_username_assign").send_keys("user2")
     self.driver.find_element(By.ID, "assignUserBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "notify_email_role")))
@@ -83,47 +94,45 @@ class Test_02_Add_packs:
     self.driver.find_element(By.ID, "user_role").find_element(By.XPATH, "(descendant::option)[. = 'PackView']").click()
     self.driver.find_element(By.ID, "assignDagRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td//a[contains(.,'user2')]")))
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"action=myprojects\"]").click()
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"logout=1\"]").click()
-    self.driver.find_element(By.ID, "username").send_keys("user")
-    self.driver.find_element(By.ID, "password").send_keys("abc123")
-    self.driver.find_element(By.ID, "login_btn").click()
-    self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
+    time.sleep(3)
+    self.vars["username"] = "user1"
+    sub=Sub1();sub.driver=self.driver;sub.vars=self.vars;sub.test_fn_switchuser() # Run fn switchuser
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs2\"]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs2\"]").click()
+    self.driver.execute_script("$('[name=\"id\"],[name=\"value\"]').css('max-width','250px')")
     self.driver.find_element(By.NAME, "id").send_keys("1")
     self.driver.execute_script("$('#south').remove()")
     self.driver.find_element(By.CSS_SELECTOR, "#singlepack input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
+    self.driver.execute_script("$('[name=\"id\"],[name=\"value\"]').css('max-width','250px')")
     assert len(self.driver.find_elements(By.CSS_SELECTOR, ".mod-packmgmt-okmsg")) == 0
+    self.driver.execute_script("//SETDESC:Assert packs not added due to duplicate ID")
+    self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-errmsg").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.NAME, "id").send_keys("4")
     self.driver.execute_script("$('#south').remove()")
     self.driver.find_element(By.CSS_SELECTOR, "#singlepack input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
+    self.driver.execute_script("$('[name=\"id\"],[name=\"value\"]').css('max-width','250px')")
     assert len(self.driver.find_elements(By.CSS_SELECTOR, ".mod-packmgmt-okmsg")) > 0
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"action=myprojects\"]").click()
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"logout=1\"]").click()
-    self.driver.find_element(By.ID, "username").send_keys("user2")
-    self.driver.find_element(By.ID, "password").send_keys("abc123")
-    self.driver.find_element(By.ID, "login_btn").click()
-    self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
+    self.driver.execute_script("//SETDESC:Assert packs added successfully")
+    self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-okmsg").send_keys("SAVESCREENSHOT")
+    self.vars["username"] = "user2"
+    sub=Sub1();sub.driver=self.driver;sub.vars=self.vars;sub.test_fn_switchuser() # Run fn switchuser
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs2\"]")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs2\"]")) == 0
+    self.driver.execute_script("//SETDESC:Assert cannot access add packs page")
+    self.driver.find_element(By.XPATH, "//table[@class='mod-packmgmt-listtable']//tr[contains(.,'packs2')]//td[3]").send_keys("SAVESCREENSHOT")
     self.driver.execute_script("window.location.href=$('a[href*=\"page=packs_list\"][href*=\"cat_id=packs2\"]').attr('href').replace('page=packs_list','page=packs_add')")
     None if len(elements := self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs2\"]")) == 0 else WebDriverWait(self.driver, 30).until(expected_conditions.staleness_of(elements[0]))
     assert len(self.driver.find_elements(By.XPATH, "//p[contains(.,'Add Packs')]")) == 0
     self.driver.execute_script("window.history.back()")
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"action=myprojects\"]").click()
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"logout=1\"]").click()
-    self.driver.find_element(By.ID, "username").send_keys("admin")
-    self.driver.find_element(By.ID, "password").send_keys("abc123")
-    self.driver.find_element(By.ID, "login_btn").click()
-    self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
+    self.vars["username"] = "admin"
+    sub=Sub1();sub.driver=self.driver;sub.vars=self.vars;sub.test_fn_switchuser() # Run fn switchuser
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"UserRights/index.php\"]").click()
     self.driver.find_element(By.XPATH, "//*[@id='user_rights_roles_table']//td//a[contains(.,'user2')]").click()
     self.driver.find_element(By.CSS_SELECTOR, "#tooltipBtnRemoveProject button").click()
