+++ tests/test_09_Assign_pack_minimization.py
@@ -28,11 +28,13 @@ class Test_09_Assign_pack_minimization:
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     time.sleep(2)
     self.driver.execute_script("$('#south').remove()")
-    self.driver.execute_script("$('input[name=\"cat_id\"]').val('packs9')")
+    self.driver.execute_script("$('[name=\"cat_id\"]').css('max-width','200px')")
+    self.driver.find_element(By.NAME, "cat_id").send_keys("packs9")
     self.driver.find_element(By.CSS_SELECTOR, "input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     time.sleep(2)
     self.driver.execute_script("$('#south').remove()")
+    self.driver.execute_script("$('[name=\"logic\"],[name=\"packfield\"],[name=\"datefield\"],[name=\"countfield\"],[name=\"valuefield\"],[name=\"roles_view\"],[name=\"roles_dags\"],[name=\"roles_invalid\"],[name=\"roles_assign\"],[name=\"roles_add\"],[name=\"roles_edit\"]').css('max-width','250px')")
     self.driver.find_element(By.NAME, "enabled").find_element(By.CSS_SELECTOR, "*[value='1']").click()
     self.driver.find_element(By.NAME, "trigger").find_element(By.CSS_SELECTOR, "*[value='M']").click()
     self.driver.find_element(By.NAME, "nominim").find_element(By.CSS_SELECTOR, "*[value='S']").click()
@@ -49,19 +51,26 @@ class Test_09_Assign_pack_minimization:
     self.driver.find_element(By.CSS_SELECTOR, "#catform button.btn-primaryrc").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=configure_edit\"][href*=\"cat_id=packs9\"]")) > 0
+    self.driver.execute_script("//SETDESC:Assert category is saved")
+    self.driver.find_element(By.XPATH, "//td//span[contains(.,'packs9')]").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs9\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href=\"#multiplepacks\"]").click()
     self.driver.execute_script("fd=new FormData($('form[enctype=\"multipart/form-data\"]')[0]);fd.set('packs_upload',new Blob([decodeURIComponent('id,value%0A1,A%0A2,B%0A3,B')]));fetch( window.location.href, {body:fd, method:'post'})")
+    self.driver.execute_script("//SAVEDESC:Perform upload of CSV data")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=configure\"]").click()
     self.driver.execute_script("$.get(window.location.href+'&runtest=assignMinimPack&record_id=1&list_minim_codes=%5B%22A%22%2C%22B%22%5D&minim_field=',function(data){$('body').attr('data-result',data)})")
+    self.driver.execute_script("//SAVEDESC:Simulate minimization")
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "body[data-result]")))
     self.driver.execute_script("res=JSON.parse($('body').attr('data-result'));if(res.packID == '1' && res.randoCode == 'A'){$('body').attr('data-minim','1')}")
+    self.driver.execute_script("//SAVEDESC:Verify minimization result")
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "body[data-minim]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=configure\"]:not([href*=\"logout=1\"])").click()
     self.driver.execute_script("$.get(window.location.href+'&runtest=assignMinimPack&record_id=1&list_minim_codes=%5B%22A%22%2C%22B%22%5D&minim_field=',function(data){$('body').attr('data-result',data)})")
+    self.driver.execute_script("//SAVEDESC:Simulate minimization")
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "body[data-result]")))
     self.driver.execute_script("res=JSON.parse($('body').attr('data-result'));if(res.randoCode == 'B'){$('body').attr('data-minim','1')}")
+    self.driver.execute_script("//SAVEDESC:Verify minimization result")
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "body[data-minim]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=configure_edit\"][href*=\"cat_id=packs9\"]").click()
     self.driver.execute_script("$('#south').remove()")
@@ -69,6 +78,8 @@ class Test_09_Assign_pack_minimization:
     self.driver.find_element(By.CSS_SELECTOR, "#catform button.btn-primaryrc").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     self.driver.execute_script("$.get(window.location.href+'&runtest=assignMinimPack&record_id=1&list_minim_codes=%5B%22A%22%2C%22B%22%5D&minim_field=',function(data){$('body').attr('data-result',data)})")
+    self.driver.execute_script("//SAVEDESC:Simulate minimization (allocation does not have a pack available)")
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "body[data-result]")))
     self.driver.execute_script("res=JSON.parse($('body').attr('data-result'));if(res === false){$('body').attr('data-minim','1')}")
+    self.driver.execute_script("//SAVEDESC:Verify minimization failure")
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "body[data-minim]")) > 0
