+++ tests/test_13_Edit_and_delete_packs.py
@@ -9,6 +9,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
+from fn_switchuser import Test_fn_switchuser as Sub1
 
 class Test_13_Edit_and_delete_packs:
   def setup_method(self, method):
@@ -19,7 +20,6 @@ class Test_13_Edit_and_delete_packs:
 
   def test_13_Edit_and_delete_packs(self):
     self.driver.get("http://127.0.0.1/")
-    self.driver.execute_script("sessionStorage.setItem('github_js_actions','1')")
     self.driver.find_element(By.LINK_TEXT, "My Projects").click()
     assert len(self.driver.find_elements(By.XPATH, "//*[@id=\"table-proj_table\"][contains(.,'Pack Management Test')]")) > 0
     self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
@@ -29,11 +29,13 @@ class Test_13_Edit_and_delete_packs:
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     time.sleep(2)
     self.driver.execute_script("$('#south').remove()")
-    self.driver.execute_script("$('input[name=\"cat_id\"]').val('packs13')")
+    self.driver.execute_script("$('[name=\"cat_id\"]').css('max-width','200px')")
+    self.driver.find_element(By.NAME, "cat_id").send_keys("packs13")
     self.driver.find_element(By.CSS_SELECTOR, "input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     time.sleep(2)
     self.driver.execute_script("$('#south').remove()")
+    self.driver.execute_script("$('[name=\"logic\"],[name=\"packfield\"],[name=\"datefield\"],[name=\"countfield\"],[name=\"valuefield\"],[name=\"roles_view\"],[name=\"roles_dags\"],[name=\"roles_invalid\"],[name=\"roles_assign\"],[name=\"roles_add\"],[name=\"roles_edit\"]').css('max-width','250px')")
     self.driver.find_element(By.NAME, "enabled").find_element(By.CSS_SELECTOR, "*[value='1']").click()
     self.driver.find_element(By.NAME, "trigger").find_element(By.CSS_SELECTOR, "*[value='F']").click()
     self.driver.find_element(By.NAME, "form").find_element(By.CSS_SELECTOR, "*[value='visit_lab_data']").click()
@@ -52,22 +54,27 @@ class Test_13_Edit_and_delete_packs:
     self.driver.find_element(By.CSS_SELECTOR, "#catform button.btn-primaryrc").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=configure_edit\"][href*=\"cat_id=packs13\"]")) > 0
+    self.driver.execute_script("//SETDESC:Assert category is saved")
+    self.driver.find_element(By.XPATH, "//td//span[contains(.,'packs13')]").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_add\"][href*=\"cat_id=packs13\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href=\"#multiplepacks\"]").click()
     self.driver.execute_script("fd=new FormData($('form[enctype=\"multipart/form-data\"]')[0]);fd.set('packs_upload',new Blob([decodeURIComponent('id,f_test%0A1,11%0A2,22%0A3,33')]));fetch( window.location.href, {body:fd, method:'post'})")
+    self.driver.execute_script("//SAVEDESC:Perform upload of CSV data")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]:not([href*=\"logout\"])").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs13\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=2\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"f_test\"][value=\"22\"]")) > 0
-    self.driver.execute_script("$('input[name=\"f_test\"]').val('1234')")
+    self.driver.find_element(By.NAME, "f_test").send_keys("1234")
     self.driver.execute_script("$('#south').remove()")
     time.sleep(0.5)
     self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-formtable input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, ".mod-packmgmt-okmsg")) > 0
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"f_test\"][value=\"1234\"]")) > 0
+    self.driver.execute_script("//SETDESC:Assert pack updated")
+    self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-okmsg").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs13\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=3\"]").click()
     self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-formtable input[type=\"button\"]").click()
@@ -75,27 +82,27 @@ class Test_13_Edit_and_delete_packs:
     self.driver.find_element(By.CSS_SELECTOR, "[role=\"dialog\"] .ok-button").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=3\"]")) == 0
+    self.driver.execute_script("//SETDESC:Assert pack ID 3 not present")
+    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=2\"]").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"UserRights/index.php\"]").click()
-    self.driver.find_element(By.ID, "new_username_assign").send_keys("user")
+    self.driver.find_element(By.ID, "new_username_assign").send_keys("user1")
     self.driver.find_element(By.ID, "assignUserBtn").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "notify_email_role")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.visibility_of_element_located((By.ID, "notify_email_role")))
     self.driver.execute_script("$('#notify_email_role').prop('checked',false)")
     self.driver.find_element(By.ID, "user_role").find_element(By.XPATH, "(descendant::option)[. = 'PackEdit']").click()
     self.driver.find_element(By.ID, "assignDagRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td//a[contains(.,'user')]")))
+    time.sleep(3)
     self.driver.find_element(By.ID, "new_username_assign").send_keys("user2")
     self.driver.find_element(By.ID, "assignUserBtn").click()
-    WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "notify_email_role")))
+    WebDriverWait(self.driver, 30).until(expected_conditions.visibility_of_element_located((By.ID, "notify_email_role")))
     self.driver.execute_script("$('#notify_email_role').prop('checked',false)")
     self.driver.find_element(By.ID, "user_role").find_element(By.XPATH, "(descendant::option)[. = 'PackView']").click()
     self.driver.find_element(By.ID, "assignDagRoleBtn").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.XPATH, "//*[@id='user_rights_roles_table']//td//a[contains(.,'user2')]")))
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"action=myprojects\"]").click()
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"logout=1\"]").click()
-    self.driver.find_element(By.ID, "username").send_keys("user")
-    self.driver.find_element(By.ID, "password").send_keys("abc123")
-    self.driver.find_element(By.ID, "login_btn").click()
-    self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
+    time.sleep(3)
+    self.vars["username"] = "user1"
+    sub=Sub1();sub.driver=self.driver;sub.vars=self.vars;sub.test_fn_switchuser() # Run fn switchuser
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs13\"]")) > 0
@@ -103,35 +110,33 @@ class Test_13_Edit_and_delete_packs:
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=2\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"f_test\"][value=\"1234\"]")) > 0
-    self.driver.execute_script("$('input[name=\"f_test\"]').val('22')")
+    self.driver.find_element(By.NAME, "f_test").send_keys("22")
     self.driver.execute_script("$('#south').remove()")
     time.sleep(0.5)
     self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-formtable input[type=\"submit\"]").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, ".mod-packmgmt-okmsg")) > 0
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "input[name=\"f_test\"][value=\"22\"]")) > 0
+    self.driver.execute_script("//SETDESC:Assert pack updated")
+    self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-okmsg").send_keys("SAVESCREENSHOT")
     self.driver.find_element(By.CSS_SELECTOR, ".mod-packmgmt-formtable input[type=\"button\"]").click()
     self.driver.execute_script("$('#south').remove()")
     self.driver.find_element(By.CSS_SELECTOR, "[role=\"dialog\"] .ok-button").click()
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.ID, "south")))
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=2\"]")) == 0
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"action=myprojects\"]").click()
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"logout=1\"]").click()
-    self.driver.find_element(By.ID, "username").send_keys("user2")
-    self.driver.find_element(By.ID, "password").send_keys("abc123")
-    self.driver.find_element(By.ID, "login_btn").click()
-    self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
+    self.driver.execute_script("//SETDESC:Assert pack ID 2 not present")
+    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"][href*=\"pack_id=1\"]").send_keys("SAVESCREENSHOT")
+    self.vars["username"] = "user2"
+    sub=Sub1();sub.driver=self.driver;sub.vars=self.vars;sub.test_fn_switchuser() # Run fn switchuser
     WebDriverWait(self.driver, 30).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]")))
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"prefix=pack_management\"][href*=\"page=packs\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs13\"]")) > 0
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"page=packs_list\"][href*=\"cat_id=packs13\"]").click()
     assert len(self.driver.find_elements(By.CSS_SELECTOR, "a[href*=\"page=packs_edit\"]")) == 0
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"action=myprojects\"]").click()
-    self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"logout=1\"]").click()
-    self.driver.find_element(By.ID, "username").send_keys("admin")
-    self.driver.find_element(By.ID, "password").send_keys("abc123")
-    self.driver.find_element(By.ID, "login_btn").click()
-    self.driver.find_element(By.LINK_TEXT, "Pack Management Test").click()
+    self.driver.execute_script("//SETDESC:Assert option to edit pack not present")
+    self.driver.find_element(By.CSS_SELECTOR, "input[data-pack-chkbx=\"1\"]").send_keys("SAVESCREENSHOT")
+    self.vars["username"] = "admin"
+    sub=Sub1();sub.driver=self.driver;sub.vars=self.vars;sub.test_fn_switchuser() # Run fn switchuser
     self.driver.find_element(By.CSS_SELECTOR, "a[href*=\"UserRights/index.php\"]").click()
     self.driver.find_element(By.XPATH, "//*[@id='user_rights_roles_table']//td//a[contains(.,'user2')]").click()
     self.driver.find_element(By.CSS_SELECTOR, "#tooltipBtnRemoveProject button").click()
